requirements.txt
python-nmap
scapy
pyshark
cryptography
PyYAML
joblib
scikit-learn
fastapi
uvicorn
pandas
prometheus_client
python-jose[cryptography]
passlib[bcrypt]
SQLAlchemy
alembic
pytest